+++ example/lib/main_page.dart
@@ -1,0 +1,156 @@
+import 'package:chewie/chewie.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:video_player/video_player.dart';
+
+class MainPage extends StatelessWidget {
+  const MainPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('主页面'),
+        actions: <Widget>[
+          TextButton(
+            onPressed: () => _goSubPage(context),
+            child: const Text(
+              '进入子页面',
+              style: TextStyle(color: Colors.white),
+            ),
+          ),
+        ],
+      ),
+      body: SingleChildScrollView(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: const <Widget>[
+            _SVG(),
+            _Video(),
+            _Toast(),
+          ],
+        ),
+      ),
+    );
+  }
+
+  void _goSubPage(BuildContext context) {
+    Navigator.of(context).pushNamed('/sub');
+  }
+}
+
+class _Video extends StatefulWidget {
+  const _Video({Key? key}) : super(key: key);
+
+  @override
+  State<_Video> createState() => _VideoState();
+}
+
+class _VideoState extends State<_Video> {
+  late VideoPlayerController _videoPlayerController;
+  late ChewieController _chewieController;
+
+  @override
+  void initState() {
+    super.initState();
+    _initVideo();
+  }
+
+  @override
+  void dispose() {
+    _videoPlayerController.dispose();
+    _chewieController.dispose();
+    super.dispose();
+  }
+
+  Future<void> _initVideo() async {
+    _videoPlayerController = VideoPlayerController.network(
+        'https://flutter.github.io/assets-for-api-docs/assets/videos/butterfly.mp4');
+    _chewieController = ChewieController(
+      videoPlayerController: _videoPlayerController,
+      autoInitialize: true,
+      aspectRatio: 16 / 9,
+      autoPlay: true,
+      looping: true,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Theme(
+      data: Theme.of(context).copyWith(platform: TargetPlatform.iOS),
+      child: AspectRatio(
+        aspectRatio: _chewieController.aspectRatio ?? (16 / 9),
+        child: Chewie(controller: _chewieController),
+      ),
+    );
+  }
+}
+
+class _SVG extends StatelessWidget {
+  const _SVG({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        Row(
+          children: <Widget>[
+            Expanded(
+              child: SvgPicture.asset(
+                'assets/svg/dart.svg',
+                placeholderBuilder: (BuildContext context) => Container(
+                  padding: const EdgeInsets.all(30.0),
+                  child: const CircularProgressIndicator(),
+                ),
+              ),
+            ),
+            Expanded(
+              child: SvgPicture.asset(
+                'assets/svg/flutter_logo.svg',
+                placeholderBuilder: (BuildContext context) => Container(
+                  padding: const EdgeInsets.all(30.0),
+                  child: const CircularProgressIndicator(),
+                ),
+              ),
+            ),
+          ],
+        ),
+        const Text(
+          'this is svg',
+          style: TextStyle(color: Colors.redAccent),
+        ),
+      ],
+    );
+  }
+}
+
+class _Toast extends StatelessWidget {
+  const _Toast({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 16),
+      child: ElevatedButton(
+        onPressed: _showToast,
+        child: const Text('Fluttertoast'),
+      ),
+    );
+  }
+
+  void _showToast() {
+    Fluttertoast.showToast(
+      msg: "This is Center Short Toast",
+      toastLength: Toast.LENGTH_SHORT,
+      gravity: ToastGravity.CENTER,
+      timeInSecForIosWeb: 1,
+      backgroundColor: Colors.red,
+      textColor: Colors.white,
+      fontSize: 16.0,
+      webPosition: 'center',
+    );
+  }
+}