+++ lib/src/optimize/flutter.js.dart
@@ -87,7 +87,8 @@ _flutter.loader = null;
     _loadEntrypoint(entrypointUrl) {
       if (!this._scriptLoaded) {
         this._scriptLoaded = new Promise((resolve, reject) => {
-          const promises = Object.keys(jsManifest).filter(key => /main.dart_\d.js/g.test(key)).sort().map(key => `${assetBase}${jsManifest[key]}`).map(this._downloadSplitJs);
+          // The code below is injected by flutter web optimizer, do not edit!!!!!!
+          const promises = Object.keys(mainjsManifest).filter(key => /^main\.dart_(\d)\.js$/g.test(key)).sort().map(key => `${assetBase}${mainjsManifest[key]}`).map(this._downloadSplitJs);
           Promise.all(promises).then((values)=>{
             const contents = values.join("");
             const script = document.createElement("script");
@@ -109,7 +110,7 @@ _flutter.loader = null;
               element.style.display = "block";
               element.style.color = "#f89800";
               element.innerText = "加载失败，点击重新请求页面";
-              document.body.appendChild(a);
+              document.body.appendChild(element);
             } else {
               this._loadEntrypoint(entrypointUrl);
             }