+++ lib/src/optimize/flutter_web_optimizer.js.dart
@@ -1,0 +1,76 @@
+const String flutterWebOptimizerSourceCode = r'''
+    // The code below is injected by flutter web optimizer, do not edit!!!!!!
+    
+    // window.assetBase
+    var assetBase = null;
+    
+    // window.mainjsManifest
+    var mainjsManifest = null;
+    
+    // window.hashFileManifest
+    var hashFileManifest = null;
+    
+    // hook dynamic create element `src` `href` property
+    (function(){
+      var _defineProperty = function (element, property) {
+        var value;
+        Object.defineProperty(element, property, {
+          get: function () {
+            return value;
+          },
+          set: function (v) {
+            value = v;
+            if (v.startsWith('./')) {
+              v = v.replace('./', '');
+            }
+            if (hashFileManifest[v]) {
+              value = assetBase + hashFileManifest[v];
+            }
+            element.setAttribute(property, value);
+          }
+        })
+  
+        var _setAttribute = element.setAttribute;
+        element.setAttribute = function () {
+          var args = Array.prototype.slice.call(arguments);
+          // if (args[0] === 'src') {
+          //   //do something here
+          //   console.log('setAttribute', args[1])
+          // }
+          _setAttribute.apply(element, args);
+        }
+      }
+  
+      var _createElement = document.createElement;
+      var _hookProperties = {'link': 'href', 'script': 'src', 'audio': 'src', 'video': 'src', 'map': 'src', 'img': 'src'};
+      document.createElement = function (tagName) {
+        var element = _createElement.call(document, tagName);
+        var property = _hookProperties[tagName.toLowerCase()];
+        if (property) {
+          _defineProperty(element, property);
+        }
+        return element;
+      }  
+    })();
+
+    // load main.dart.js_xxx.part js
+    function dartDeferredLibraryLoader(uri, successCallback, errorCallback, loadId) {
+      console.info('uri: ' + uri + ', loadId: '+ loadId);
+      var src;
+      try {
+        var url = new URL(uri);
+        var key = url.pathname.replaceAll(/(.*)(main\.dart\.(.+)\.js)/g, '$2');
+        src = assetBase + mainjsManifest[key];
+      } catch (e) {
+        var key = uri.replaceAll(/(.*)(main\.dart\.(.+)\.js)/g, '$2');
+        src = assetBase + mainjsManifest[key];
+      }
+    
+      var script = document.createElement("script");
+      script.type = "text/javascript";
+      script.src = src;
+      script.addEventListener("load", successCallback, false);
+      script.addEventListener("error", errorCallback, false);
+      document.body.appendChild(script);
+    }
+    ''';