+++ lib/src/optimize/optimize_command.dart
@@ -7,14 +7,15 @@ import 'dart:typed_data';
 
 import 'package:args/command_runner.dart';
 import 'package:crypto/crypto.dart' as crypto;
+import 'package:flutter_web_optimizer/src/optimize/flutter_service_worker.js.dart';
+import 'package:flutter_web_optimizer/src/optimize/flutter_web_optimizer.js.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart' show parse;
+import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as path;
 
 import '../common/logger.dart';
 import '../common/model.dart';
-import 'dart_agency_build.dart';
-import 'dart_deferred_library_loader.js.dart';
 import 'flutter.js.dart';
 
 class OptimizeCommand extends Command<void> {
@@ -37,9 +38,10 @@ class OptimizeCommand extends Command<void> {
             'only support relative path，root path is [path.context.current]，'
             'eq：flutter_web_optimize_plugin.dart',
       )
-      ..addOption(
-        'target',
-        help: 'Build target'
+      ..addFlag(
+        'enable-pwa',
+        help: 'enable PWA service worker',
+        defaultsTo: true,
       );
   }
 
@@ -51,25 +53,13 @@ class OptimizeCommand extends Command<void> {
   String get name => 'optimize';
 
   /// 资源路径，一般是cdn地址
-  String get _assetBase => argResults!['asset-base'] ?? '';
+  late String _assetBase;
 
   /// Web构建产物路径
-  String get _webOutput {
-    final String? webOutput = argResults!['web-output'];
-    if (webOutput?.isNotEmpty ?? false) {
-      return path.join(path.context.current, webOutput!);
-    }
-    return path.join(path.context.current, 'build/web');
-  }
+  late String _webOutput;
 
   /// plugin 文件路径，支持处理资源上传cdn等操作
-  String get _plugin {
-    final String? pluginFile = argResults!['plugin'];
-    if (pluginFile?.isNotEmpty ?? false) {
-      return path.join(path.context.current, pluginFile!);
-    }
-    return '';
-  }
+  late String _plugin;
 
   /// isolate通信，发送信息
   SendPort? _sendPort;
@@ -80,17 +70,19 @@ class OptimizeCommand extends Command<void> {
   /// isolate消息
   Stream<IsolateMessageProtocol>? _message;
 
-  /// 哈希化后的main.dart.js && main.dart.xxx.part文件
-  final Map<String, String> _jsManifest = <String, String>{};
+  // 需要hash的文件
+  final Map<String, String> _hashFiles = <String, String>{};
 
-  /// 需要上传的文件
-  List<String> _toUploadFiles = <String>[];
+  // 哈希化后的文件清单
+  final Map<String, String> _hashFileManifest = <String, String>{};
 
   @override
   FutureOr<void> run() async {
     Logger.info('start web optimize inspireui');
 
-    await _initIsolate(argResults!.arguments);
+    await _parseArgs();
+
+    await _initIsolate();
 
     final Directory outputDir = Directory(_webOutput);
     if (!outputDir.existsSync()) {
@@ -99,20 +91,24 @@ class OptimizeCommand extends Command<void> {
       exit(2);
     }
 
-    final File mainDartJsFile = outputDir
-        .listSync()
-        .whereType<File>()
-        .singleWhere(
-            (File entity) => path.basename(entity.path) == 'main.dart.js');
-    await _splitMainDartJS(mainDartJsFile);
-
-    _hashScripts();
-
-    await _cdnAssets();
+    await _splitMainDartJS(outputDir);
 
     _replaceFlutterJS();
 
+    // _hashAssets();
+
+    _importFlutterWebOptimizerJS();
+
+    _updateSourceMapsMark();
+
+    final bool enablePWA = argResults!['enable-pwa'];
+    if (enablePWA) {
+      _updateFlutterServiceWorkerJS();
+    }
+
     _injectToHtml();
+
+    await _cdnAssets();
 
     _disposeIsolate();
 
@@ -152,61 +148,91 @@ class OptimizeCommand extends Command<void> {
   //   }
   // }
 
-  /// 解析 flutter_web_optimizer.yaml 文件
-  // void _parseYaml() {
-  //   final String contents =
-  //       File('${path.current}/flutter_web_optimizer.yaml').readAsStringSync();
-  //   final dynamic doc = loadYaml(contents);
-  //   if (doc is! Map) {
-  //     return;
-  //   }
-  //   _assetBase = doc['asset_base'] ?? '';
-  //   if (_assetBase.isEmpty) {
-  //     throw Exception(
-  //         '[flutter_web_optimizer.yaml]: asset_base_url args not be null!!!');
-  //   }
-  //   if (doc['cdn_upload'] is! Map<dynamic, dynamic>) {
-  //     return;
-  //   }
-  //   if (doc['cdn_upload']['custom_upload_file_name'] is! String) {
-  //     return;
-  //   }
-  //   _plugin = doc['cdn_upload']['custom_upload_file_name'] ?? '';
-  // }
+  Future<void> _parseArgs() async {
+    /// 资源路径，一般是cdn地址
+    _assetBase = argResults!['asset-base'] ?? '';
+
+    /// Web构建产物路径
+    final String? webOutput = argResults!['web-output'];
+    if (webOutput?.isNotEmpty ?? false) {
+      _webOutput = path.join(path.context.current, webOutput!);
+    } else {
+      _webOutput = path.join(path.context.current, 'build', 'web');
+    }
+
+    /// plugin 文件路径，支持处理资源上传cdn等操作
+    final String? plugin = argResults!['plugin'];
+    if (plugin?.isEmpty ?? true) {
+      _plugin = '';
+      return;
+    }
+    if (path.extension(plugin!).isNotEmpty) {
+      /// Uri
+      _plugin = path.join(path.context.current, plugin);
+    } else {
+      /// Not Uri
+      final PackageConfig? packageConfig =
+          await findPackageConfig(Directory.current);
+      if (packageConfig == null) {
+        _plugin = '';
+        return;
+      }
+      try {
+        final Package package = packageConfig.packages
+            .singleWhere((Package element) => element.name == plugin);
+        _plugin = path.join(package.packageUriRoot.toFilePath(), plugin);
+        _plugin = path.setExtension(_plugin, '.dart');
+      } catch (_) {
+        _plugin = '';
+      }
+    }
+    if (!File(_plugin).existsSync()) {
+      throw Exception('plugin args is invalid!!!');
+    }
+  }
 
   /// 初始化isolate通信
-  Future<void> _initIsolate(List<String> args) {
-    final Completer<void> completer = Completer<void>();
+  Future<void> _initIsolate() {
     if (_plugin.isEmpty) {
-      completer.complete();
+      return Future<void>.value();
     }
+
+    final Completer<void> completer = Completer<void>();
     final StreamController<IsolateMessageProtocol> controller =
         StreamController<IsolateMessageProtocol>.broadcast();
     _message = controller.stream;
 
     _receivePort ??= ReceivePort();
     _receivePort!.listen((dynamic message) {
-      Logger.info('server isolate get message: $message');
-      if (message is Map<String, Object>) {
-        controller.add(IsolateMessageProtocol.fromMap(message));
-      }
       if (message is SendPort) {
         _sendPort = message;
         completer.complete();
       }
+
+      if (message is Map<String, Object>) {
+        Logger.info('server isolate get message: $message');
+        controller.add(IsolateMessageProtocol.fromMap(message));
+      }
     });
 
-    Isolate.spawnUri(Uri.file(_plugin), args, _receivePort!.sendPort);
+    Isolate.spawnUri(
+      Uri.file(_plugin),
+      argResults!.arguments,
+      _receivePort!.sendPort,
+    );
     return completer.future;
   }
 
   /// 释放isolate通信
   void _disposeIsolate() {
     _receivePort?.close();
+    _receivePort = null;
+    _sendPort = null;
+    _message = null;
   }
 
   /// 拆分 main.dart.js
-  Future<void> _splitMainDartJS(File file) async {
+  Future<void> _splitMainDartJS(Directory outputDir) async {
     // 写入单个文件
     Future<bool> writeSingleFile({
       required File file,
@@ -239,9 +265,16 @@ class OptimizeCommand extends Command<void> {
       return completer.future;
     }
 
+    final File file = outputDir.listSync().whereType<File>().singleWhere(
+        (File entity) => path.basename(entity.path) == 'main.dart.js');
+
+    /// 针对xhr加载动态js文件，插入『@ sourceURL』标记，方便调试
+    file
+        .openSync(mode: FileMode.append)
+        .writeStringSync('\n\n//@ sourceURL=main.dart.js\n');
     const int totalChunk = 6;
     final Uint8List bytes = file.readAsBytesSync();
-    int chunkSize = (bytes.length / totalChunk).ceil();
+    final int chunkSize = (bytes.length / totalChunk).ceil();
     final List<Future<bool>> futures = List<Future<bool>>.generate(
       totalChunk,
       (int index) {
@@ -255,132 +288,392 @@ class OptimizeCommand extends Command<void> {
     );
 
     await Future.wait(futures);
-
-    /// 分片完成后删除 main.dart.js
-    file.deleteSync();
-  }
-
-  /// md5文件
-  String _md5File(File file, {String? md5}) {
-    final Uint8List bytes = file.readAsBytesSync();
-    // 截取8位即可
-    final md5Hash = md5 ?? crypto.md5.convert(bytes).toString().substring(0, 8);
-
-    // 文件名使用hash值
-    final basename = path.basenameWithoutExtension(file.path);
-    final extension = path.extension(file.path);
-    return '$md5Hash-$basename$extension';
-  }
-
-  void _hashScripts() {
-    var md5 = crypto.md5
-        .convert([
-          Random().nextInt(1000),
-          Random().nextInt(1000),
-          Random().nextInt(1000),
-          Random().nextInt(1000)
-        ])
-        .toString()
-        .substring(0, 8);
-    Logger.info('hash script $md5');
-    Directory(_webOutput)
-        .listSync()
-        .whereType<File>()
-        .where(
-            (File file) => RegExp(r'main\.dart(.*)\.js$').hasMatch(file.path))
-        .forEach((File file) {
-      _toUploadFiles.add(file.path);
-      final String key = path.relative(file.path, from: _webOutput);
-      final String filename = _md5File(file, md5: md5);
-      file.renameSync(path.join(path.dirname(file.path), filename));
-      if (key.startsWith('main.dart_')) {
-        _jsManifest[path.basename(file.path)] = filename;
-      }
-    });
-
-    /// rewrite to main script
-    for (var mainItem in _jsManifest.keys.toList()) {
-      final File file = File('$_webOutput/${_jsManifest[mainItem]}');
-      if (file.existsSync()) {
-        String contents = file.readAsStringSync();
-        contents = contents.replaceAll(
-            RegExp(r'main.dart.js'), '$md5-main.dart.js');
-        file.writeAsString(contents);
-      } else {
-        Logger.info(file.path);
-      }
-    }
-    Logger.info('hash script done');
-  }
-
-  /// 资源cdn化
-  Future<void> _cdnAssets() async {
-    final Completer<void> completer = Completer<void>();
-    if (_plugin.isEmpty) {
-      completer.complete();
-    }
-    _message?.listen((IsolateMessageProtocol protocol) {
-      if (protocol.isResponse &&
-          protocol.action == IsolateMessageAction.cdnAssets) {
-        completer.complete();
-      }
-    });
-    _sendPort?.send(IsolateMessageProtocol.request(
-      IsolateMessageAction.cdnAssets,
-      _toUploadFiles,
-    ).toMap());
-    return completer.future;
   }
 
   /// 替换 flutter.js
   void _replaceFlutterJS() {
-    File('$_webOutput/flutter.js').writeAsStringSync(flutterJSSourceCode);
+    final File file = File('$_webOutput/flutter.js');
+    if (!file.existsSync()) {
+      file.createSync();
+    }
+    file.writeAsStringSync(flutterJSSourceCode);
   }
 
-  /// 向 index.html 注入
-  void _injectToHtml() {
-    final String? target = argResults?['target'];
-    final bool isAgency = target == 'agency';
-    final File file = File('$_webOutput/index.html');
-    String contents = file.readAsStringSync();
-    final File flutterJsFile = File('$_webOutput/flutter.js');
-    final String filename = _md5File(flutterJsFile);
-    flutterJsFile
-        .renameSync(path.join(path.dirname(flutterJsFile.path), filename));
-    contents = contents.replaceAll(
-      RegExp(r'<script src="flutter.js" defer></script>'),
-      '<script src="$filename" defer></script>',
-    );
-    final Document document = parse(contents);
-    final String dartDeferredLibraryLoader = dartDeferredLibraryLoaderSourceCode
-        .replaceAll(
-            RegExp('var assetBase = null;'), 'var assetBase = "$_assetBase";')
-        .replaceAll(
-          RegExp('var jsManifest = null;'),
-          'var jsManifest = ${jsonEncode(_jsManifest)};',
-        );
-    final List<Element> scripts = document.getElementsByTagName('script');
-    bool isInjected = false;
-    for (int i = 0; i < scripts.length; i++) {
-      final Element element = scripts[i];
-      if (element.text.contains(RegExp(r'var serviceWorkerVersion'))) {
-        element.text = '${element.text}\n$dartDeferredLibraryLoader';
-        if (isAgency) {
-          element.text = '${element.text}\n$dartAgencyBuild';
-        }
-        isInjected = true;
-        break;
-      }
+  /// md5文件
+  String _md5File(File file) {
+    Logger.info('hashing file: ${file.path}');
+    final Uint8List bytes = file.readAsBytesSync();
+    // 截取8位即可
+    final String md5Hash = crypto.md5.convert(bytes).toString().substring(0, 8);
+
+    // 文件名使用hash值
+    final String basename = path.basenameWithoutExtension(file.path);
+    final String extension = path.extension(file.path);
+    final String filename = '$basename.$md5Hash$extension';
+    Logger.info('hashed file name: $filename');
+    return filename;
+  }
+
+  /// 资源hash化
+  void _hashAssets() {
+    /// 替换
+    String replace(
+      Match match,
+      File file,
+      String key,
+      Map<String, String> hashFiles,
+    ) {
+      // 文件名使用hash值
+      final String filename = _md5File(file);
+      // 此时key已经是 posix 平台的分隔符 /
+      // 统一将路径分隔符修改成 / ，assetManifest 和 fontManifest 使用的是 posix 平台的分隔符 /
+      final String dirname = path.posix.dirname(key);
+      final String newKey = path.posix.join(dirname, filename);
+
+      // hash文件路径
+      final String newPath = path.join(path.dirname(file.path), filename);
+      hashFiles[file.path] = newPath;
+
+      return '${match[1]}$newKey${match[3]}';
     }
-    if (!isInjected) {
-      final Element? headElement = document.head;
+
+    /// hash化icons目录
+    void hashIconsDir() {
+      // 读取 manifest.json 清单文件
+      final File manifest = File(path.join(_webOutput, 'manifest.json'));
+      String manifestContents = manifest.readAsStringSync();
+      // 遍历构建产物icons目录
+      final Directory iconsDir = Directory(path.join(_webOutput, 'icons'));
+      iconsDir
+          .listSync()
+          .whereType<File>() // 文件类型
+          .where((File file) => !path.basename(file.path).startsWith('.'))
+          .forEach((File file) {
+        // 使用 posix 平台的分隔符 /
+        String key = path.relative(file.path, from: _webOutput);
+        key = path.toUri(key).toString();
+        // 替换 manifest.json 清单文件
+        manifestContents = manifestContents.replaceFirstMapped(
+          RegExp('(.*)($key)(.*)'),
+          (Match match) => replace(match, file, key, _hashFiles),
+        );
+      });
+      // 写入修改后的 manifest.json 清单文件
+      manifest.writeAsStringSync(manifestContents);
+    }
+
+    /// hash化assets目录
+    void hashAssetsDir(Map<String, File> files) {
+      // 读取资源清单文件
+      final File assetManifest =
+          File(path.join(_webOutput, 'assets', 'AssetManifest.json'));
+      String assetManifestContents = assetManifest.readAsStringSync();
+      // 读取字体清单文件
+      final File fontManifest =
+          File(path.join(_webOutput, 'assets', 'FontManifest.json'));
+      String fontManifestContents = fontManifest.readAsStringSync();
+
+      // 遍历构建产物assets目录，对资源文件md5后获取哈希值，并修改资源、字体清单文件
+      final Directory assetsDir = Directory(path.join(_webOutput, 'assets'));
+      assetsDir
+          .listSync(recursive: true)
+          .whereType<File>() // 文件类型
+          .where((File file) => !path.basename(file.path).startsWith('.'))
+          .forEach((File file) {
+        // 使用 posix 平台的分隔符 /
+        String key = path.relative(file.path, from: assetsDir.path);
+        key = path.toUri(key).toString();
+        // 替换资源清单文件
+        assetManifestContents = assetManifestContents.replaceFirstMapped(
+          RegExp('(.*)($key)(.*)'),
+          (Match match) => replace(match, file, key, _hashFiles),
+        );
+        // 替换字体清单文件
+        fontManifestContents = fontManifestContents.replaceAllMapped(
+          RegExp('(.*)($key)(.*)'),
+          (Match match) => replace(match, file, key, _hashFiles),
+        );
+
+        if (<String>['.svg'].contains(path.extension(file.path))) {
+          // 针对flutter_svg插件使用svg文件的特殊处理
+          files[key] = file;
+        }
+      });
+
+      // 写入修改后的资源、字体清单文件
+      assetManifest.writeAsStringSync(assetManifestContents);
+      fontManifest.writeAsStringSync(fontManifestContents);
+    }
+
+    /// 修正源码引用
+    void amendSourceCode(Map<String, File> files) {
+      // 将AssetManifest.json文件进行md5
+      String assetManifestFileName = 'AssetManifest.json';
+      final File assetManifestFile =
+          File(path.join(_webOutput, 'assets', assetManifestFileName));
+      files[assetManifestFileName] = assetManifestFile;
+
+      // 将FontManifest.json进行md5
+      String fontManifestFileName = 'FontManifest.json';
+      final File fontManifestFile =
+          File(path.join(_webOutput, 'assets', fontManifestFileName));
+      files[fontManifestFileName] = fontManifestFile;
+
+      // 需要修正源码中关联的字符串引用
+      final Map<String, String> amendSourceCodeRelatedStrings =
+          <String, String>{};
+      files.forEach((String key, File value) {
+        String newKey = key.replaceFirstMapped(
+          RegExp('(.*)($key)(.*)'),
+          (Match match) => replace(match, value, key, _hashFiles),
+        );
+        if (newKey.startsWith('./')) {
+          newKey = newKey.substring(2);
+        }
+        amendSourceCodeRelatedStrings[key] = newKey;
+      });
+
+      Directory(_webOutput).listSync().whereType<File>() // 文件类型
+          .where((File file) {
+        final RegExp regExp = RegExp(r'^main\.dart(.*)\.js$');
+        final String filename = path.basename(file.path);
+        return regExp.hasMatch(filename);
+      }).forEach((File file) {
+        // 查找 main.dart_xxx.js 和 main.dart.xxx.part.js 文件
+        // 修正修正源码中关联的字符串引用
+        String contents = file.readAsStringSync();
+        amendSourceCodeRelatedStrings.forEach((String key, String value) {
+          contents = contents.replaceAll(RegExp(key), value);
+        });
+        file.writeAsStringSync(contents);
+      });
+    }
+
+    /// hash化web根目录
+    void hashWebOutputDir() {
+      Directory(_webOutput).listSync().whereType<File>() // 文件类型
+          .where((File file) {
+        final RegExp regExp = RegExp(
+            r'(main\.dart(.*)\.js)|(favicon.png)|(flutter.js)|(manifest.json)');
+        final String filename = path.basename(file.path);
+        return regExp.hasMatch(filename);
+      }).forEach((File file) {
+        final String filename = _md5File(file);
+        _hashFiles[file.path] = path.join(path.dirname(file.path), filename);
+      });
+    }
+
+    // 需要修正源码中关联的文件引用
+    final Map<String, File> amendSourceCodeRelatedFiles = <String, File>{};
+
+    hashIconsDir();
+    hashAssetsDir(amendSourceCodeRelatedFiles);
+    amendSourceCode(amendSourceCodeRelatedFiles);
+    hashWebOutputDir();
+
+    // 重命名文件
+    _hashFiles.forEach((String key, String value) {
+      File(key).renameSync(value);
+    });
+  }
+
+  /// 导入flutter_web_optimizer.js
+  void _importFlutterWebOptimizerJS() {
+    /// 排序后生成 _hashFileManifest 和 jsManifest
+    void getManifestValue(Map<String, String> jsManifest) {
+      _hashFiles.keys.toList()
+        ..sort()
+        ..forEach((String key) {
+          String filepath =
+              path.toUri(path.relative(key, from: _webOutput)).toString();
+          String hashedFilePath = path
+              .toUri(path.relative(_hashFiles[key]!, from: _webOutput))
+              .toString();
+          _hashFileManifest[filepath] = hashedFilePath;
+          final String filename = path.basename(key);
+          if (RegExp(r'^main\.dart(.*)\.js$').hasMatch(filename)) {
+            jsManifest[filename] = path.basename(_hashFiles[key]!);
+          }
+        });
+    }
+
+    /// 修正flutter_web_optimizer.js源码
+    File amendSourceCode(Map<String, String> jsManifest) {
+      // 更新 assetBase、mainjsManifest 和 hashFileManifest
+      const JsonEncoder jsonEncoder = JsonEncoder.withIndent('  ');
+      final String flutterWebOptimizer = flutterWebOptimizerSourceCode
+          .replaceAll(
+            RegExp('var assetBase = null;'),
+            'var assetBase = "$_assetBase";',
+          )
+          .replaceAll(
+            RegExp('var mainjsManifest = null;'),
+            'var mainjsManifest = ${jsonEncoder.convert(jsManifest)};',
+          )
+          .replaceAll(
+            RegExp('var hashFileManifest = null;'),
+            'var hashFileManifest = ${jsonEncoder.convert(_hashFileManifest)};',
+          );
+      final File file = File('$_webOutput/flutter_web_optimizer.js')
+        ..createSync()
+        ..writeAsStringSync(flutterWebOptimizer);
+      return file;
+    }
+
+    // 哈希化后的main.dart.js && main.dart.xxx.part文件
+    final Map<String, String> jsManifest = <String, String>{};
+    getManifestValue(jsManifest);
+    final File file = amendSourceCode(jsManifest);
+
+    // 哈希flutter_web_optimizer.js文件
+    final String fileName = _md5File(file);
+    String filepath = file.path;
+    String hashedFilePath = path.join(path.dirname(file.path), fileName);
+    _hashFiles[filepath] = hashedFilePath;
+    // 重命名文件
+    file.renameSync(hashedFilePath);
+    // 更新 _hashFileManifest
+    filepath = path.toUri(path.relative(filepath, from: _webOutput)).toString();
+    hashedFilePath =
+        path.toUri(path.relative(hashedFilePath, from: _webOutput)).toString();
+    _hashFileManifest[filepath] = hashedFilePath;
+  }
+
+  /// 修正//# sourceMappingURL，插入//@ sourceURL
+  void _updateSourceMapsMark() {
+    String replace(Match m) => '${m[1]}$_assetBase${_hashFileManifest[m[2]]}';
+    Directory(_webOutput).listSync().whereType<File>() // 文件类型
+        .where((File file) {
+      final RegExp regExp = RegExp(r'^main\.dart(.*)\.js$');
+      final String filename = path.basename(file.path);
+      return regExp.hasMatch(filename);
+    }).forEach((File file) {
+      String contents = file.readAsStringSync();
+      contents = contents
+          .replaceFirstMapped(RegExp(r'(//# sourceMappingURL=)(.*)\n'), replace)
+          .replaceFirstMapped(RegExp(r'(//@ sourceURL=)(.*)\n'), replace);
+      file.writeAsStringSync(contents);
+    });
+  }
+
+  /// 更新flutter_service_worker.js
+  void _updateFlutterServiceWorkerJS() {
+    final File file = File(path.join(_webOutput, 'flutter_service_worker.js'));
+    if (!file.existsSync()) {
+      return;
+    }
+    String contents = file.readAsStringSync();
+    if (contents.isEmpty) {
+      return;
+    }
+    String? resourcesValues = RegExp(r'const RESOURCES = {\n([\s\S]*)\n};')
+        .firstMatch(contents)
+        ?.group(1);
+    if (resourcesValues?.isEmpty ?? true) {
+      return;
+    }
+    String? coreValues = RegExp(r'const CORE = \[\n([\s\S]*?)\];')
+        .firstMatch(contents)
+        ?.group(1);
+    if (coreValues?.isEmpty ?? true) {
+      return;
+    }
+
+    final List<String> insertedJSList = <String>[];
+    _hashFileManifest.keys.where((String key) => key != 'main.dart.js').toList()
+      ..sort()
+      ..forEach((String key) {
+        final String value = _hashFileManifest[key]!;
+        resourcesValues =
+            resourcesValues?.replaceAll(RegExp('"$key"'), '"$value"');
+        coreValues = coreValues?.replaceAll(RegExp('"$key"'), '"$value"');
+
+        final String filename = path.basename(key);
+        if (RegExp(r'^(main\.dart(.*)\.js|flutter_web_optimizer\.js)$')
+            .hasMatch(filename)) {
+          // 收集main.dart.*js和flutter_web_optimizer.js文件
+          insertedJSList.add(value);
+        }
+      });
+    resourcesValues = resourcesValues?.replaceAll(
+        RegExp(r'"main\.dart\.js": (.+),'),
+        insertedJSList
+            .where((String key) => RegExp(
+                    r'^main\.dart_\d(.*)\.js|flutter_web_optimizer\.(.*)\.js$')
+                .hasMatch(key))
+            .map((String js) {
+              final String md5String = crypto.md5
+                  .convert(File(path.join(_webOutput, js)).readAsBytesSync())
+                  .toString();
+              return '"$js": "$md5String",';
+            })
+            .toList()
+            .join('\n'));
+    coreValues = coreValues?.replaceAll(
+      RegExp(r'"main\.dart\.js",'),
+      insertedJSList
+          .where((String key) =>
+              !RegExp(r'^flutter_web_optimizer\.(.*)\.js$').hasMatch(key))
+          .map((String js) => '"$js",')
+          .toList()
+          .join('\n'),
+    );
+    contents = flutterServiceWorkerSourceCode
+        .replaceFirstMapped(
+          RegExp(r'(const RESOURCES = {)(};)'),
+          (m) => '${m[1]}\n$resourcesValues\n${m[2]}',
+        )
+        .replaceAll(
+          RegExp(RegExp.escape('const CORE = [];'), caseSensitive: true),
+          'const CORE = [\n$coreValues\n];',
+        )
+        .replaceAll(
+          RegExp(r'const assetBase = null;'),
+          'var assetBase = "$_assetBase";',
+        );
+    file.writeAsStringSync(contents);
+  }
+
+  /// 修改 index.html
+  void _injectToHtml() {
+    /// 注入 assetBase meta 标签
+    void injectAssetBaseMeta({
+      required Document document,
+      required Element? headElement,
+    }) {
       if (headElement != null) {
-        final Element script = Element.tag('script');
-        script.text = '\n$dartDeferredLibraryLoader';
-        if (isAgency) {
-          script.text = '${script.text}\n$dartAgencyBuild';
+        void injectAssetBaseMeta(Element element) {
+          final Element meta = Element.tag('meta');
+          meta.attributes['name'] = 'assetBase';
+          meta.attributes['content'] = _assetBase;
+
+          element.append(Text('\n'));
+          element.append(Comment('content值必须以 / 结尾'));
+          element.append(Text('\n'));
+          element.append(meta);
+          element.append(Text('\n'));
         }
 
+        final List<Element> metas = document.getElementsByTagName('meta');
+        if (metas.isNotEmpty) {
+          injectAssetBaseMeta(metas.last);
+        } else {
+          injectAssetBaseMeta(headElement);
+        }
+      }
+    }
+
+    /// 注入 flutter_web_optimizer.js script
+    void injectFlutterWebOptimizerScript({
+      required Document document,
+      required Element? headElement,
+    }) {
+      if (headElement != null) {
+        final Element script = Element.tag('script');
+        script.attributes['src'] = 'flutter_web_optimizer.js';
+        script.attributes['defer'] = '';
+
+        final List<Element> scripts = document.getElementsByTagName('script');
         if (scripts.length > 1) {
           final Element firstScript = scripts.first;
           headElement.insertBefore(script, firstScript);
@@ -391,6 +684,78 @@ class OptimizeCommand extends Command<void> {
         }
       }
     }
+
+    /// 替换哈希后的文件
+    String replace(String key, {bool isAbsolute = true}) {
+      if (key.isEmpty || !_hashFileManifest.containsKey(key)) {
+        return key;
+      }
+      if (isAbsolute) {
+        return '$_assetBase${_hashFileManifest[key]}';
+      } else {
+        return _hashFileManifest[key]!;
+      }
+    }
+
+    /// 读取index.html
+    final File file = File('$_webOutput/index.html');
+    String contents = file.readAsStringSync();
+
+    /// 解析 index.html 为 Document
+    final Document document = parse(contents);
+
+    final Element? headElement = document.head;
+
+    injectAssetBaseMeta(document: document, headElement: headElement);
+    injectFlutterWebOptimizerScript(
+      document: document,
+      headElement: headElement,
+    );
+
+    // 查找所有元素
+    final List<Element> elements = document.querySelectorAll('*');
+    for (final Element element in elements) {
+      if (element.attributes.containsKey('href')) {
+        String href = element.attributes['href']!;
+        final bool isAbsolute = !<String>[
+          'icons/Icon-192.png',
+          'favicon.png',
+          'manifest.json',
+        ].contains(href);
+        href = replace(href, isAbsolute: isAbsolute);
+        element.attributes['href'] = href;
+      }
+      if (element.attributes.containsKey('src')) {
+        String src = element.attributes['src']!;
+        src = replace(src);
+        element.attributes['src'] = src;
+      }
+    }
+
+    // 写入文件
     file.writeAsStringSync(document.outerHtml);
+  }
+
+  /// 资源cdn化
+  Future<void> _cdnAssets() async {
+    if (_plugin.isEmpty) {
+      return Future<void>.value();
+    }
+    final Completer<void> completer = Completer<void>();
+    StreamSubscription<IsolateMessageProtocol>? subscription;
+    subscription = _message?.listen((IsolateMessageProtocol protocol) {
+      if (protocol.isResponse &&
+          protocol.action == IsolateMessageAction.cdnAssets) {
+        subscription?.cancel();
+        completer.complete();
+      }
+    });
+    // 需要上传的文件
+    final List<String> toUploadFiles = _hashFiles.values.toList();
+    _sendPort?.send(IsolateMessageProtocol.request(
+      IsolateMessageAction.cdnAssets,
+      toUploadFiles,
+    ).toMap());
+    return completer.future;
   }
 }